Attempted bug fix
I tried using the debugging feature for code blocks and with some time I realized a possible culprit. There is a small section that is commented out of my program at line 98 to line 102, this section seemed to trigger a segmentation fault. Once I removed it, the program still passed all other tests and the error was gone by leaving me with an "All tests completed!" rather than any kind of segmentation fault.