Staging of the out of the box Romi example program
Did remap the controller inputs for the XBox controller - it originally used axis(2) as the rotation which on the XBox controller is the analog "LT" control - changed it to use axis(0) for the turn and it worked fine.  Axis(1) was already used for forward/backward and the example code generated already had the controller negation for forward/backward done (the online WPILib doc said you had to do it - it's probably a little out of date).