WIP - testing the ramsete code
Working the trajectory to better understand what the Romi is doing.

So far we seem to read distances about 50% short and the final pose command in the trajectory seems to make the robot go nuts...

Tried messing with the sign of the Z angle and it didn't seem to help much - next up, slow the turning rate?